fedora-kata-osbuilder.sh: Don't copy kernel, use a symlink
Keeping a copy of the host kernel elsewhere is bad, as it opens up
worries that we may have outdated and insecure kernels in user on
the host. So this removes the kernel caching.

The implication is if the symlinked kernel is removed from the host
we now have a broken symlink and kata won't work until we regenerate
symlinks.

If kata-osbuilder-generate.service is enabled, it will run as oneshot
on boot, and set the symlink to point to the host kernel, which dnf
will prevent removal of while the host is using it. So if the service
is running, the kernel can never be removed from underneath us.

So this change essentially requires that service to be enabled
to ensure correct behavior on kernel removal.

Signed-off-by: Cole Robinson <crobinso@redhat.com>